\name{write.broadpeak.info}
\alias{write.broadpeak.info}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Write out determined broad enrichment regions using broadPeak format }
\description{
  Writes out broad regions of enrichment determined by the
  get.broad.enrichment.clusters method in a broadPeak format.
}
\usage{
write.broadpeak.info(broadpeak.results, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{broadpeak.results}{ output of the \code{\link{get.broad.enrichment.clusters}} }
  \item{filename}{ file name }
}
